refactor: extract cpu to separate file
